Move pCardanoEra parser to Cardano.CLI.Common.Parsers
