+++ cardano-cli/src/Cardano/CLI/Parsers.hs
@@ -6,12 +6,9 @@
 module Cardano.CLI.Parsers
   ( opts
   , pref
-  , pConsensusModeParams
-  , pNetworkId
   ) where
 
 import           Cardano.CLI.Byron.Parsers (backwardsCompatibilityCommands, parseByronCommands)
-import           Cardano.CLI.Common.Parsers (pConsensusModeParams, pNetworkId)
 import           Cardano.CLI.Environment (EnvCli)
 import           Cardano.CLI.Ping (parsePingCmd)
 import           Cardano.CLI.Render (customRenderHelp)
